websocket

+++ Comet-API/src/main/java/com/cometproject/api/networking/messages/IMessageEvent.java
@@ -15,7 +15,7 @@ public interface IMessageEvent {
 
     byte[] toRawBytes();
 
-    int getId();
+    short getId();
 
     int getLength();
 }
